[1] Support rendering 10/16-bit YUV into Bgra64 bitmap buffer. [2] Support operating with Bgra64 bitmap buffer by Session correctly.

+++ PixelViewer/Media/ImageRenderers/P210ImageRenderer.cs
@@ -16,7 +16,7 @@ namespace Carina.PixelViewer.Media.ImageRenderers
 
 
         // Select UV component.
-        protected override void SelectUV(byte uv1, byte uv2, out byte u, out byte v)
+        protected override void SelectUV(ushort uv1, ushort uv2, out ushort u, out ushort v)
         {
             u = uv1;
             v = uv2;
